services/analytics-dashboard/requirements.txt
dash==2.18.2
dash-bootstrap-components==1.6.0
plotly==5.24.1
matplotlib==3.9.3
pandas==2.2.3
numpy==2.1.3
gunicorn==23.0.0
flask-cors==5.0.0
python-dotenv==1.0.1
